bug#7681

+++ src/views/RecommendProductsList.vue
@@ -42,7 +42,7 @@
   </div>
 </template>
 <script>
-import { successNotify, errorNotify } from '@shared/util'
+import { successNotify, errorNotify, confirmBox } from '@shared/util'
 import RecommondUrl from '@api/recommendProducts/recommendProductsUrl.js'
 import RecommondApi from '@api/recommendProducts/recommendProducts.js'
 import CommonApi from '@api/api.js'
@@ -184,40 +184,50 @@ export default {
       const PUSHPRODUCTS = SELECTIONARR && SELECTIONARR.length > 0 ? SELECTIONARR : [row.id]
       // 批量推品供方货号和单独供方货号
       const ITEMNOALL = SELECTIONARR && SELECTIONARR.length > 0 ? ITEMNOALLARR : row.supplierItemNo
-      RecommondApi.recommend({ productIdList: PUSHPRODUCTS })
+      confirmBox(this, '是否提交商品', '')
         .then(() => {
-          successNotify(this, `供方货号：${ITEMNOALL}提交成功`)
-          this.$refs.listView.refresh()
-        })
-        .catch(() => {
-          errorNotify(this, `供方货号：${ITEMNOALL}提交失败`)
+          RecommondApi.recommend({ productIdList: PUSHPRODUCTS })
+            .then(() => {
+              successNotify(this, `供方货号：${ITEMNOALL}提交成功`)
+              this.$refs.listView.refresh()
+            })
+            .catch(() => {
+              errorNotify(this, `供方货号：${ITEMNOALL}提交失败`)
+            })
         })
     },
     deleteProduct (row) {
-      RecommondApi.deleteRecommed(row.id)
-        .then(res => {
-          this.gotoPage()
-          successNotify(this, `供方货号：${row.supplierItemNo}删除成功`)
-        })
-        .catch(res => {
-          errorNotify(this, `供方货号：${row.supplierItemNo}删除失败`)
+      confirmBox(this, '是否删除商品', '')
+        .then(() => {
+          RecommondApi.deleteRecommed(row.id)
+            .then(res => {
+              this.gotoPage()
+              successNotify(this, `供方货号：${row.supplierItemNo}删除成功`)
+            })
+            .catch(res => {
+              errorNotify(this, `供方货号：${row.supplierItemNo}删除失败`)
+            })
         })
     },
     cancel (row) {
-      RecommondApi.cancelrcommend(row.id)
-        .then(res => {
-          this.$refs.listView.refresh()
-          successNotify(this, `供方货号：${row.supplierItemNo}取消成功`)
-        })
-        .catch(() => {
-          errorNotify(this, `供方货号：${row.supplierItemNo}取消失败`)
+      confirmBox(this, '是否取消提交', '')
+        .then(() => {
+          RecommondApi.cancelrcommend(row.id)
+            .then(res => {
+              this.$refs.listView.refresh()
+              successNotify(this, `供方货号：${row.supplierItemNo}取消提交成功`)
+            })
+            .catch(() => {
+              errorNotify(this, `供方货号：${row.supplierItemNo}取消提交失败`)
+            })
         })
     },
     OdmDetail (status, row) {
+      const { id, categoryId, supplierItemNo } = row
       if (status !== 'create') {
-        this.$router.push({ path: '/home/recommend-products/OdmDetail', query: { mode: status, id: row.id, categoryId: row.categoryId } })
+        this.$router.push({ path: '/home/recommend-products/OdmDetail', query: { mode: status, id: id, categoryId: categoryId, supplierItemNo: supplierItemNo } })
       } else {
-        this.$router.push({ path: '/home/recommend-products/OdmOneDetails', query: { categoryId: row.categoryId, mode: status, id: row.id } })
+        this.$router.push({ path: '/home/recommend-products/OdmOneDetails', query: { categoryId: categoryId, mode: status, id: id, supplierItemNo: supplierItemNo } })
       }
     }
   }
